refactor: share responsedto for controller class
